Update readme.md
Initially I wanted to make the backend on Python as I have learned a lot of things on Python in the past weeks/months (followed Linkedin courses on concurrent/parallel progr, design patterns, inner working with the GIL and a bunch of stuff in the standard library, I did some exercices on Leetcode/Hackerrank in Python). I had chosen this language to learn in depth bc I was already familiar with it, I like it, it is very fast to write python code and learning difficult concepts with a language I was already familiar with was a pretty good idea.

However most of this knowledge is transferable (most of the concepts are similar, the implementation and syntax differ and differences must be found. For instance python does not allow parallel threads while JS do, python allow multiple inheritance while JS does not as it is based on prototypes, array/string/map/file system interaction/etc... methods have different syntax...) as I have not yet dig into frameworks and became proficient with them, nor did I build a real world project.

As I want to learn frontend dev, learning JS was necessary and as the majority of job offers I have found are in nodejs making that switch is a bit of a "sacrifice" but for a greater good (my favorite companies work also on this stack). I also found a training course for an apprenticeship in JavaScript so I had all the reason to make that difficult change in my decision.

But I still like python and I will probably do side projects with it. It is a versatile language that can be used for AI/ML, software / web programming, a bit of game development and maybe one day it will be of  used in another context and the investment was worth for learning new concepts. 

Regex was totally transferable mais bref